refactor(apm): Move structs into implementing files
