+++ pkg/apm/deployments.go
@@ -4,6 +4,22 @@ import (
 	"fmt"
 )
 
+// Deployment represents information about a New Relic application deployment.
+type Deployment struct {
+	Links       *DeploymentLinks `json:"links,omitempty"`
+	ID          int              `json:"id,omitempty"`
+	Revision    string           `json:"revision"`
+	Changelog   string           `json:"changelog,omitempty"`
+	Description string           `json:"description,omitempty"`
+	User        string           `json:"user,omitempty"`
+	Timestamp   string           `json:"timestamp,omitempty"`
+}
+
+// DeploymentLinks contain the application ID for the deployment.
+type DeploymentLinks struct {
+	ApplicationID int `json:"application,omitempty"`
+}
+
 // ListDeployments returns deployments for an application.
 func (apm *APM) ListDeployments(applicationID int) ([]*Deployment, error) {
 	deployments := []*Deployment{}
